chore: move services to the domain purchase root
